v 7756

+++ Shared/A2v10.Request/Models/ShellModel.cs
@@ -15,7 +15,7 @@ namespace A2v10.Request.Models
 	public class SwitchToCompanySaveModel
 	{
 		public Int64 UserId { get; set; }
-		public Int64 TenantId { get; set; }
+		public Int32 TenantId { get; set; }
 		public Int64 CompanyId { get; set; }
 	}
